Add ability to specify where the menu is appended
Currently the context menu DOM element (#cy-context-menus-cxt-menu) is
appended to the document's body element.

When the menu is positioned at the time of the user's click, the "left"
and "right" positions are calculated based on the rendered graph's
container's location plus the cilck position within the graph.

If the graph was rendered in a container which was moved away from the
normal flow position in the body (such as for the purpose of containing
it in a "slide- in" animation), then the calculated position for the
context menu might not be correct based on the current positioning of
elements in the browser.

To accommodate this, add an option that allows for supplying a CSS
selector that the #cy-context-menus-cxt-menu element should be appended
to instead of the body.