Case: edge context menu with an item, background context menu empty.
When show context menu over edge, show the menu with one item. When call event directly on background, without closing the menu before, show a context menu with no item, but it's possible to see the borders.

This happens because this function, that is failing in checking element visibility. Accordingly to "http://code.jquery.com/jquery-latest.js", it should check if the display style is equal 'none' instead checking if only existis. In my case, the style.display was block, and it was returning true for isElementHidden.